update deployment manifest
I'm storing the port information in a configMap because I want to be able to reference it in other manifests. This way if I ever want to update it, I'll be able to change it in just two locations, the configserver deployment itself, and the configmap. Then all other references to it will be pointing to the config map. Or at least that's the plan.